+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Removed
+* Resolving ambiguity using positional distance
 
 ## [1.12](https://pypi.org/project/multimethod/1.12/) - 2024-07-04
 ### Changed